requirements.txt
psutil==5.9.5
keyboard==0.13.5
pywin32==306
pyautogui==0.9.54
Pillow==10.0.0
pystray==0.19.4
